+++ clients/src/main/java/org/apache/kafka/server/telemetry/ClientTelemetryExporter.java
