+++ clients/src/main/java/org/apache/kafka/server/telemetry/ClientTelemetryReceiver.java
@@ -23,11 +23,11 @@ import org.apache.kafka.server.authorizer.AuthorizableRequestContext;
  * {@code ClientTelemetryReceiver} defines the behaviour for telemetry receiver on the broker side
  * which receives client telemetry metrics.
  *
- * @deprecated Since 4.1.0, use {@link ClientTelemetryExporter} instead. This interface will be
+ * @deprecated Since 4.2.0, use {@link ClientTelemetryExporter} instead. This interface will be
  *             removed in Kafka 5.0.0. The new interface provides additional context including
  *             the push interval to enable better metric lifecycle management.
  */
-@Deprecated
+@Deprecated(since = "4.2", forRemoval = true)
 public interface ClientTelemetryReceiver {
     /**
      * Called by the broker when a client reports telemetry metrics. The associated request context
